to_do_app/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.6
django-bootstrap3==10.0.0
django-crispy-forms==1.12.0
django-crum==0.7.9
django-filter==2.4.0
django-floppyforms==1.9.0
django-tables2==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
fontawesome-free==5.15.4
future==0.18.2
html5lib==1.1
idna==3.3
numpy==1.21.2
pandas==1.3.2
pdfkit==0.6.1
Pillow==8.3.1
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.61.0
urllib3==1.26.7
webencodings==0.5.1
xhtml2pdf==0.2.5
